pyproject.toml
[project]
name = "end-to-end-cardest-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "tqdm>=4.67.1",
]
